Fixed move bug
