+++ tests/conftest.py
@@ -12,7 +12,7 @@ from confluent_kafka.cimpl import Producer, TopicPartition
 from pykafka.exceptions import NoBrokersAvailableError
 
 from esque.cluster import Cluster
-from esque.config import Config, sample_config_path
+from esque.config.config import Config, sample_config_path
 from esque.errors import raise_for_kafka_error
 from esque.resources.consumergroup import ConsumerGroupController
 from esque.resources.topic import Topic
